refactor: view, domain, infrastructure로 패키지 이동
