+++ src/main/java/step2/domain/NumberGenerateStrategy.java
@@ -1,7 +1,8 @@
-package step2;
+package step2.domain;
 
 import java.util.List;
 
+@FunctionalInterface
 public interface NumberGenerateStrategy {
     List<Integer> generate();
 }
