+++ src/main/java/step2/view/InputView.java
@@ -1,9 +1,8 @@
-package step2;
+package step2.view;
 
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
-import java.util.stream.Collector;
 import java.util.stream.Collectors;
 
 public class InputView {
